Ok great! Now to improve the UX to match the flow of the in-game contract process, I'd like to update the Add Contract Item form. Maybe just make a button to "Add Contract" and then have the form be in a modal?
It should allow the user to input ALL the destinations for a contract as well as what and how many SCU of each good. Eg., "X amount of Y good delivered to Z location".

Then as more contracts are entered, any similar locations/destinations can be updated like they currently are