Ok great, it's scrolling now!
Now can we add a second tab called Log Book? I'd like the completed contracts to be placed here. To do this we may need to edit or add some data structures though because currently we are treating each "destination" as a contract, but that's not the case.

When a location is marked complete, it should just be used for the player to check it off their list and update the cargo totals. Once all destinations for a contract have been marked complete, THEN that contract can be considered complete. We'll likely need to track this when a contract is added since all of the destinations will be added at that time.

Does this make sense or do you need any additional context or clarification?